# docs/swagger.yaml
basePath: /
definitions:
...
  requests.FinishRunningRequest:
    properties:
      latitude:
        example: 35.6815
        type: number
      longitude:
        example: 139.7675
        type: number
    type: object
  requests.GPSPointRequest:
    properties:
      accuracy:
        example: 5
        type: number
      client_id:
        description: PWA側で生成したユニークID（オプション）
        example: uuid-v4-here
        type: string
      latitude:
        example: 35.6812
        type: number
      longitude:
        example: 139.7671
        type: number
      timestamp:
        example: "2026-02-10T07:01:00Z"
        type: string
    type: object
...
  requests.SendGPSPointsRequest:
    properties:
      points:
        items:
          $ref: '#/definitions/requests.GPSPointRequest'
        type: array
    type: object
  requests.StartRunningRequest:
    properties:
      latitude:
        example: 35.6812362
        type: number
      longitude:
        example: 139.7671248
        type: number
    type: object
  response.ActivityResponse:
    properties:
      auto_detected:
        example: false
        type: boolean
      created_at:
        example: "2026-02-10T07:00:00Z"
        type: string
      distance_km:
        example: 5.234
        type: number
      duration_min:
        example: 35
        type: integer
      ended_at:
        type: string
      exercise_type:
        example: running
        type: string
      gps_points:
        items:
          $ref: '#/definitions/response.GPSPointResponse'
        type: array
      gym_location_id:
        type: string
      gym_location_name:
        type: string
      id:
        example: 01JARQ3KEXAMPLE00003
        type: string
      started_at:
        example: "2026-02-10T07:00:00Z"
        type: string
      status:
        example: in_progress
        type: string
      team_id:
        example: 01JARQ3KEXAMPLE00001
        type: string
      updated_at:
        example: "2026-02-10T07:00:00Z"
        type: string
      user_id:
        example: firebaseUID123
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        example: invalid_token
        type: string
      message:
        example: エラーが発生しました
        type: string
    type: object
  response.GPSPointResponse:
    properties:
      accuracy:
        example: 5
        type: number
      latitude:
        example: 35.6812
        type: number
      longitude:
        example: 139.7671
        type: number
      timestamp:
        example: "2026-02-10T07:00:00Z"
        type: string
    type: object
...
  response.SendGPSPointsResponse:
    properties:
      current_distance_km:
        example: 3.456
        type: number
      saved_count:
        example: 2
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Trihackathon API Server
  title: Trihackathon API
  version: "1.0"
paths:
  /api/activities:
    get:
      description: 自分のアクティビティ一覧を取得する
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.ActivityResponse'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 自分のアクティビティ一覧
      tags:
      - activities
...
  /api/activities/running/{activityId}:
    get:
      description: 指定したランニングアクティビティの詳細情報（GPSポイント含む）を取得する
      parameters:
      - description: アクティビティID
        in: path
        name: activityId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ActivityResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ランニング記録詳細
      tags:
      - activities-running
  /api/activities/running/{activityId}/finish:
    post:
      consumes:
      - application/json
      description: ランニングアクティビティを完了する。GPSポイントから総移動距離を再計算しdistance_kmを確定。
      parameters:
      - description: アクティビティID
        in: path
        name: activityId
        required: true
        type: string
      - description: 終了地点情報
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.FinishRunningRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ActivityResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ランニング完了
      tags:
      - activities-running
  /api/activities/running/{activityId}/gps:
    post:
      consumes:
      - application/json
      description: バックグラウンドで蓄積したGPSデータをバッチ送信する。精度が50mを超えるポイントは距離計算から除外（保存はする）。
      parameters:
      - description: アクティビティID
        in: path
        name: activityId
        required: true
        type: string
      - description: GPSポイントデータ
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.SendGPSPointsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SendGPSPointsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: GPSポイント送信（バッチ）
      tags:
      - activities-running
  /api/activities/running/start:
    post:
      consumes:
      - application/json
      description: ランニングアクティビティを開始する。同時に進行中にできるアクティビティは1つのみ。チームがactive状態かつexercise_typeがrunningの場合のみ。
      parameters:
      - description: 開始地点情報
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.StartRunningRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ActivityResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ランニング開始
      tags:
      - activities-running
...
  /api/teams/{teamId}/activities:
    get:
      description: チーム全体のアクティビティ一覧を取得する
      parameters:
      - description: チームID
        in: path
        name: teamId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.ActivityResponse'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: チーム全体のアクティビティ一覧
      tags:
      - activities
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"